refactor(controller): extract model storage config to model_storage.go
Part 3/7 of splitting the god-controller into per-concern files.
Tracks #311.

Move the model storage wiring — which decides between mounting a
pre-staged PVC, using a shared cache PVC, or downloading into an
ephemeral emptyDir — out of inferenceservice_controller.go into a
focused model_storage.go sibling. This is the single largest chunk
in the split (~260 lines).

Moved symbols:
  const ModelCachePVCName
  isLocalModelSource
  buildModelInitCommand
  modelInitEnvVars
  type modelStorageConfig
  buildModelStorageConfig
  buildPVCStorageConfig
  buildCachedStorageConfig
  buildEmptyDirStorageConfig
  (*InferenceServiceReconciler).ensureModelCachePVC

Pure relocation — no signature changes, no behavior changes. Callers
in the same package continue to resolve the unqualified names.

const PhaseWaitingForGPU stays in the controller file until C5
(scheduling extraction) claims it.

Line deltas (based on main; if C1 #312 and C2 #313 merge first, the
controller file drops another ~200 lines on rebase):
  inferenceservice_controller.go: 1,567 → 1,311 (−256)
  model_storage.go:                       0 → 296 (new)

Verified: go vet clean, golangci-lint 0 issues, make test green,
internal/controller coverage unchanged at 83.1%.
